Add King piece functionality
Add methods and atributes to board to know if white or black king is on the board, functionality of the first piece of each team being a king piece, updating the board's state on king pieces, and method to render the crown over king pieces. Add Crown.png texture. Update heuristic of GameEvaluator to consider the king pieces. Update GameTree class method apply_play so that it also carries over the king status of a piece to the game scenarios boards.